Fix GzDecoder Write partial filenames and comments
If the gzip header contains an optional filename or comment but they are
not completely contained in the buffer sent to a `write::GzDecoder`,
then a valid header is created, missing data from these optional
sections. A subsequent write call will treat the remaining header as
encoded data and attempt to decode it, generally causing a panic.

This change rewrites the header parsing code to handle partial headers
correctly for both `Read` (where `WouldBlock` is handled specially) and
`Write` (where `UnexpectedEof` is handled specially).