Check minimal version of Rust that compiles
Specify a minimum version of the Rust compiler that successfully
builds the crate. This allows us to be aware if the minimal version
increases and to avoid issues such as
https://github.com/rust-lang/flate2-rs/issues/370

The minimum version is a best-effort measured value and is different to
the MSRV which is a support guarantee. The minimum version can be
incremented by a PR in order to pass tests, as long as the MSRV holds.
When the minimum version increases, the next release should be a minor
version, to allow any affected users to pin to a previous minor version.